util.c: to_utf8 returns off_t instead of size_t

+++ src/internal.h
@@ -35,6 +35,6 @@ off_t parse_hex(uint64_t *dest, const char *src, size_t len);
 
 off_t parse_number(Nson *nson, const char *p, size_t len);
 
-size_t to_utf8(char *dest, const uint64_t chr, const size_t len);
+off_t to_utf8(char *dest, const uint64_t chr, const size_t len);
 
 #endif /* !INTERNAL_H */
